requirements.txt
altgraph==0.17.2
appdirs==1.4.4
asgiref==3.4.1
attrs==20.3.0
beautifulsoup4==4.11.1
cached-property==1.5.2
certifi==2020.11.8
chardet==3.0.4
defusedxml==0.6.0
Django==4.0
django-tempus-dominus==5.1.2.13
django-view-table==1.0.1
django-widget-tweaks==1.4.8
future==0.18.2
idna==2.10
isodate==0.6.0
lxml==4.6.1
mssql-django==1.1.1
numpy==1.22.3
pandas==1.4.2
pefile==2021.9.3
psycopg2-binary==2.8.6
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
#pyodbc==4.0.32
python-dateutil==2.8.2
requests==2.25.0
requests-file==1.5.1
requests-toolbelt==0.9.1
schedule==1.1.0
six==1.15.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
tzdata==2021.5
urllib3==1.26.2
zeep==4.0.0
